Add user authorization

+++ app/Http/Controllers/TweetsController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Models\User;
 use App\Models\Tweet;
 use App\Http\Resources\Tweet as TweetResource;
 
@@ -14,13 +13,13 @@ class TweetsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function getOne($id)
+    public function getFromUser($id)
     {
-      $tweet = Tweet::find($id);
-      if(!$tweet){
+      $tweet = Tweet::where('author_id',$id)->get();
+      if(count($tweet) == 0){
         return get_error(404);
       }
-      return new TweetResource($tweet);
+      return TweetResource::collection($tweet);
     }
 
     /**
@@ -45,7 +44,7 @@ class TweetsController extends Controller
     public function store(Request $request)
     {
       $tweet = new Tweet;
-      $tweet->author_id = $request->input('author_id');
+      $tweet->author_id = auth('api')->user()->id;
       $tweet->text = $request->input('text');
 
       if($tweet->save()){
@@ -67,6 +66,7 @@ class TweetsController extends Controller
         if(!$tweet){
           return get_error(404);
         }
+
         $tweet->text = $request->input('text');
 
         if($tweet->save()){
@@ -87,6 +87,11 @@ class TweetsController extends Controller
         if(!$tweet){
           return get_error(404);
         }
+        if(!isCorrectUserApi($tweet->author_id)){
+          return get_error(401);
+        }
         $tweet->delete();
+        return 204;
     }
+
 }
