+++ app/Http/Controllers/UsersController.php
@@ -4,59 +4,15 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
-use App\Models\Tweet;
 use App\Http\Resources\User as UserResource;
 use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
 
 class UsersController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function login(Request $request)
-    {
-      $user = User::where('username', $request->input('username'));
-      if(!$user){
-        return get_error(404);
-      }
-      if (Hash::check($request->input('password'), $user->password)) {
-        // The passwords match, do login
-        Auth::login($user);
-        return new UserResource($user);
-      }
-      return new get_error(401, 'Passwords doesn\'t match');
-    }
 
     /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function register(Request $request)
-    {
-      if($this->usernameAlreadyInUse){
-        return get_error(400, 'Username already in use');
-      }
-      $user = new User;
-      $user->name = $request->input('name');
-      $user->email = $request->input('email');
-      $user->username = $request->input('username');
-      $user->password = Hash::make($request->input('password'));
-
-      if($user->save()){
-        // Do login
-        Auth::login($user);
-        return new UserResource($user);
-      }
-      return get_error();
-    }
-
-    /**
-     * Show the form for creating a new resource.
+     * Get information from user
      *
      * @return \Illuminate\Http\Response
      */
@@ -66,24 +22,12 @@ class UsersController extends Controller
       if(!$user){
         return get_error(404);
       }
-      //return response()->json([$user->tweet]);
       return new UserResource($user);
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function getAll()
-    {
-      $users = User::get();
-      return UserResource::collection($users);
-    }
-
 
     /**
-     * Update the specified resource in storage.
+     * Update user name and password.
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
@@ -91,32 +35,78 @@ class UsersController extends Controller
      */
     public function update(Request $request, $id)
     {
-      $user = User::findOrFail($id);
+      $user = User::findOrFail(auth('api')->user()->id);
+
+      //Check if user exists/Authorization token is correct
       if(!$user){
         return get_error(404);
       }
-      if($request->input('email')) $user->email = $request->input('email');
-      if($request->input('username')) $user->username = $request->input('username');
-      if($request->input('password')) $user->password = Hash::make($request->input('password'));
 
-      if($user->save()){
-        return new UserResource($user);
+      //Check if user updating is the same user logged in
+      if(!isCorrectUserApi($id)){
+        return get_error(401);
       }
-      return get_error();
+
+      //Check if all the user information is entered correct
+      $validator = $this->updateValidator($request->all());
+      if($validator->fails() OR (!Hash::check($request['current_password'], $user->password))){
+          return get_error(400, $validator->errors()->all());
+      }
+
+      $user->name = $request['name'];
+      $user->password = Hash::make($request['password']);
+      $user->save();
+
+      return new UserResource($user);
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Remove user and tweets related to him.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-      $user = User::findOrFail($id);
+      $user = User::findOrFail(auth('api')->user()->id);
+
+      //Check if user exists
       if(!$user){
         return get_error(404);
       }
+
+      //Check if user deleting is the same user logged in
+      if(!isCorrectUserApi($id)){
+        return get_error(401);
+      }
+
+      //Delete tweets from user
+      $user->tweet()->where('author_id', $id)->delete();
+
+      //Delete login token from database
+      $token = auth('api')->user()->token()->revoke();
       $user->delete();
+
+      return response()->json(204);
     }
+
+
+
+
+    /**
+     * Get a validator for an incoming update request.
+     * Must insert old\current password.
+     * Password must be at least with length 8, with a special character and match the password_confirmation input.
+     * @param  array  $data
+     * @return \Illuminate\Contracts\Validation\Validator
+     */
+    protected function updateValidator(array $data)
+    {
+        return Validator::make($data, [
+            'name' => ['required', 'string', 'max:255'],
+            'current_password' => ['required'],
+            'password' => ['required', 'min:8','regex:/^.*(?=.{3,})(?=.*[a-zA-Z])(?=.*[0-9])(?=.*[\d\X])(?=.*[@&!$#%]).*$/', 'confirmed'],
+        ]);
+    }
+
 }
